chore: move mistraltokenizer into mistral folder
