+++ tests/prompt_strategies/test_chat_templates_mistral.py
@@ -8,7 +8,7 @@ import pytest
 if TYPE_CHECKING:
     from transformers import PreTrainedTokenizer
 
-    from axolotl.utils.mistral_tokenizer import HFMistralTokenizer
+    from axolotl.utils.mistral.mistral_tokenizer import HFMistralTokenizer
 
 
 # fmt: off
